Media queries added for main content to coexist well, position of animation altered, reduced margin of main content
Media queries alter the size of the SVG so that it does not overpower the card when they are both on the same row. Position of SVG when going through animation is altered so the distance it moves along y-axis is shorter. Also reduced space between main content and title.